Fix game reset bug
  * React has a problem with variable number hooks between renders. To
fix this, we moved all the refs into one useRef, and mimicked a ref using
a simple object with {current: undefined}. This way we can pass this fake
ref to the ReactComponents to obtain the ref(for calculating position);
but at the same time the parent component is happy as the hook count is
constant between renders.
  * Also added effect that forces a new render, every time the component
is rendered with undefined holePositions. This prevents the bunny from
being painted to screen without knowledge of hole positions.