Fix pubsub bug
Earlier, events were being dispatched from the document. This is a
security risk as anyone can cheat and win. Now we still depend on the
browser to handle the events, since its APIs are very robust; but we
obscure this from the public tampering, by not exposing it within any of
the DOM.